Fix comment refactor code
