+++ lib/orca_api/orca_qkan_service.rb
@@ -11,8 +11,8 @@ module OrcaApi
   #   @see OrcaQkanService::QkanProviderListService#get
   # @!method qkan_patient_service
   #   @see OrcaQkanService::QkanPatientService#create
-  # @!method qkan_homecare_master_service
-  #   @see OrcaQkanService::QkanHomecareMasterService#get
+  # @!method get_qkan_master_code_list_service
+  #   @see OrcaQkanService::QkanMasterCodeListService#get
 
   # @!endgroup
   class OrcaQkanService < Service
@@ -20,7 +20,7 @@ module OrcaApi
       PatientPaygetService
       QkanProviderListService
       QkanPatientService
-      QkanHomecareMasterService
+      QkanMasterCodeListService
     ).each do |class_name|
       method_suffix = Client.underscore(class_name)
       require_relative "orca_qkan_service/#{method_suffix}"
